refactor: move EmptyState component
