+++ src/frontend/react_app/src/components/EmptyState.tsx
